+++ R/tbl_stack.R
@@ -71,27 +71,9 @@ tbl_stack <- function(tbls) {
     stop("Expecting 'tbls' to be a list, e.g. 'tbls = list(tbl1, tbl2)'")
   }
 
-  # checking all inputs are class tbl_uvregression, tbl_regression, tbl_summary, or tbl_merge
-  if (!map_chr(tbls, ~class(.x)[1]) %in% c(
-    "tbl_regression", "tbl_uvregression",
-    "tbl_summary", "tbl_merge"
-  ) %>% any()) {
-    stop("All objects in 'tbls' must be class 'tbl_regression',
-         'tbl_uvregression', 'tbl_summary', or 'tbl_merge'")
-  }
-
-  # printing message if stacking tbl_summary and regression object
-  if ("tbl_summary" %in% map_chr(tbls, ~class(.x)[1]) &&
-    any(c("tbl_regression", "tbl_uvregression") %in% map_chr(tbls, ~class(.x)[1]))) {
-    message(paste(
-      "You are stacking a gtsummary regression table and a summary table,",
-      "which is not recommended. Consider revising the format of your table."
-    ))
-  }
-
-  # checking if there are multiple input types
-  if (map_chr(tbls, ~class(.x)[1]) %>% unique() %>% length() > 1) {
-    message("Multiple gtsummary object classes detected. Displayed results default to first input class type.")
+  # checking all inputs are class gtsummary
+  if (!purrr::every(tbls, ~inherits(.x, "gtsummary"))) {
+    stop("All objects in 'tbls' must be class 'gtsummary'", call. = FALSE)
   }
 
   # will return call, and all arguments passed to tbl_stack